requirements.txt
fastapi
uvicorn
sentence-transformers
pdfminer.six
scikit-learn
torch
faiss-cpu
python-multipart
llama-cpp-python
